Cybersecurity ChatBot Part 1
Here the who project is put together and runs well. Fixes were made like, not allowing a user to add numbers and special keys to their name, when it asks you a question you can type 'y' and 'n' for yes or no, and all answers are stored in an array.